Copied update to other projects.
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk/ifolder@392 db2431b9-df06-0410-971c-a38ea9a75485

+++ tools/Report/Report.cs
@@ -174,7 +174,7 @@ public class Report
 		writer.WriteLine("  Total: {0}", totalCases);
 		writer.WriteLine("    Ran: {0}", totalRuns);
 		writer.WriteLine(" Passed: {0} [ {1:p1} ]", totalPasses,
-			(double)(totalRuns > 0 ? (double)totalPasses/(double)totalRuns : 0));
+			(totalRuns > 0 ? (double)totalPasses/(double)totalRuns : 0));
 		writer.WriteLine();
 		writer.WriteLine("   Time: {0} seconds", totalTime);
